first tab added automatically
After updating tab the options of setting and delete are working fine (appears and disappears in on clicks)

Parent and child deletion

Incremental tab is now saving its parents ID

Grid that is dragged in content-parent is not saving its closest parent's id

+++ ajax_removeitem.php
@@ -1,13 +1,21 @@
 <?php include_once("config.php");
 	$vdivid = $_POST['item_id'];    
 	$modulename = $_POST['module_name'];    
-		
-	$sqlexist = "delete from dznr_module_fields where divid = '".$vdivid."' and module_name = '".$modulename."' ";
-	if(mysqli_query($dbcon, $sqlexist)){
-		echo 'success';
+		// first check if the vdivid has a child if it has a child then return to the user and tell them to delete the child first then delete the parent 
+	$checkChildSql="SELECT * FROM `dznr_module_fields` WHERE `parent_grid_id`= $vdivid";
+	$res=mysqli_query($dbcon,$checkChildSql);
+	$count=mysqli_num_rows($res);
+	if($count==0){
+		$sqlexist = "delete from dznr_module_fields where divid = '".$vdivid."' and module_name = '".$modulename."' ";
+		if(mysqli_query($dbcon, $sqlexist)){
+			echo 'success';
+		}else{
+			echo 'fail';
+		}
 	}else{
-		echo 'fail';
+		echo "3";// this message represents that a child exists in the database
 	}
+	
 	
 	
 ?>
